+++ StudyLink/src/model/DOB.java
@@ -1,4 +1,4 @@
-package domainobjects;
+package model;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.text.ParseException;
