+++ StudyLink/src/model/User.java
@@ -1,4 +1,4 @@
-package domainobjects;
+package model;
 
 import java.util.UUID;
