+++ StudyLink/src/view/ForgotPassword.java
@@ -1,4 +1,4 @@
-package components;
+package view;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
