+++ StudyLink/src/view/SignUp.java
@@ -1,8 +1,8 @@
-package components;
+package view;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.*;
-import domainobjects.*;
+import model.*;
 import helpers.*;
 
 /*
